LayoutTests/fast/events/message-event-data-isolated-world-expected.txt

CONSOLE MESSAGE: Tests that MessageEvent.data works as expected from isolated worlds.
CONSOLE MESSAGE: PASS: The isolated world was able to read the MessagePort sent via postMessage()
